@@ -14,21 +14,20 @@
     "server:heroku": "NODE_ENV=prod node scripts/server.js"
   },
   "engines": {
-    "node": "8.11.1"
+    "node": "12"
   },
   "private": true,
   "dependencies": {
     "@angular-redux/store": "^9.0.0",
-    "@angular/animations": "^7.2.6",
-    "@angular/cdk": "7.2.1",
-    "@angular/common": "^7.2.6",
-    "@angular/compiler": "^7.2.6",
-    "@angular/core": "^7.2.6",
-    "@angular/forms": "^7.2.6",
-    "@angular/http": "^7.2.6",
-    "@angular/platform-browser": "^7.2.6",
-    "@angular/platform-browser-dynamic": "^7.2.6",
-    "@angular/router": "^7.2.6",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@fortawesome/angular-fontawesome": "^0.3.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.12",
     "@fortawesome/free-solid-svg-icons": "^5.6.3",
@@ -36,25 +35,25 @@
     "express": "^4.16.3",
     "font-awesome": "^4.7.0",
     "http-auth": "^3.2.3",
-    "iwerk-angular-ui": "^3.9.0",
+    "iwerk-angular-ui": "^7.0.0",
     "primeicons": "1.0.0",
-    "primeng": "7.0.4",
+    "primeng": "^8.0.0",
     "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
     "redux-observable": "^1.0.0",
-    "rxjs": "^6.3.3",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.3",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.13.2",
-    "@angular/cli": "^7.2.1",
-    "@angular/compiler-cli": "^7.2.6",
-    "@angular/language-service": "^7.2.6",
+    "@angular-devkit/build-angular": "^0.803.20",
+    "@angular/cli": "^8.3.20",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
     "@types/redux-logger": "^3.0.6",
-    "codelyzer": "4.5.0",
+    "codelyzer": "^5.2.0",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "2.0.0",
@@ -66,6 +65,6 @@
     "protractor": "~5.4.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "3.2.2"
+    "typescript": "~3.5.3"
   }
 }
