@@ -1,5 +1,5 @@
 
 export const environment: Environment = {
   production: false,
-  apiKey: 'SOME_KEY_HERE_CHECK_NEWS_API_ORG'
+  apiKey: '5c2d1b9dbee44042b0faebe1451e164d'
 };
